CastCore framework without raw representable
In our application we have a number of `NS_ENUM` and `NS_OPTION` types
that we would like to embed in `DictionaryConvertible` types. The
built-in support to serialize these in json as their raw representable
value (usually an integer) is not what we want. But since Cast's
prinicipal serialization methods are overloaded to use raw value
converters, our types can not be extended to DictionaryConvertible for a
custom representation.

To avoid bringing these extensions into scope, I moved them into a
separate `ClassRaw.swift` source and created a new framework target
`CastCore` which doesn't include them. The original `Cast` framework
works as before. Since I don't know if this is useful to others, I
haven't created osx or tv variants of the stripped down framework.
