+++ cast-script/Cast.swift
@@ -75,27 +75,6 @@ public extension JSONDictionary {
         }
         return nil
     }
-
-    public func value<A: RawRepresentable>(for key: JSONKey) -> A? where A.RawValue: JSONValue  {
-        if let raw: A.RawValue = value(for: key)  {
-            return A(rawValue: raw)
-        }
-        return nil
-    }
-
-    public func value<A: RawRepresentable>(for key: JSONKey) -> [A]? where A.RawValue: JSONValue {
-        if let rawArray: [A.RawValue] = value(for: key)  {
-            return rawArray.flatMap { A(rawValue: $0) }
-        }
-        return nil
-    }
-
-}
-
-extension RawRepresentable  {
-    public var jsonValue: Any? {
-        return self.rawValue
-    }
 }
 
 extension Dictionary: JSONDictionary, JSONValue {
@@ -229,12 +208,6 @@ extension Array where Element: JSONValue {
         }
         return nil
     }
-    public var jsonValue: Any? {
-        return self.map { $0.jsonValue }
-    }
-}
-
-extension Array where Element: RawRepresentable, Element.RawValue: JSONValue {
     public var jsonValue: Any? {
         return self.map { $0.jsonValue }
     }
