+++ cast-script/CastCore.h
@@ -1,0 +1,19 @@
+//
+//  CastCore.h
+//  CastCore
+//
+//  Created by Guy on 21/05/2016.
+//  Copyright © 2016 Houzz. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for Cast.
+FOUNDATION_EXPORT double CastVersionNumber;
+
+//! Project version string for Cast.
+FOUNDATION_EXPORT const unsigned char CastVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <Cast/PublicHeader.h>
+
+
